major refactor of the underlying architecture
mapping is now cleaner and more functional, should be easier to test individual pieces

added a few utility functions
