+++ htmap/__init__.py
@@ -1,7 +1,8 @@
 from .settings import settings
+from .mapping import map, starmap
 from .mapper import htmap, HTMapper, MapBuilder
 from .result import MapResult, JobStatus
 from .options import MapOptions
-from .shortcuts import map, starmap, build_map, recover, remove
-from .management import clean, map_ids, map_results, status
+from .shortcuts import recover, remove, force_remove
+from .management import clean, map_ids, map_results, status, force_clean
 from . import exceptions
