+++ htmap/shortcuts.py
@@ -1,18 +1,8 @@
 from typing import Callable
 
+import shutil
+
 from . import mapper, result, exceptions
-
-
-def map(map_id: str, func: Callable, args, force_overwrite: bool = False, **kwargs) -> result.MapResult:
-    return mapper.htmap(func).map(map_id, args, force_overwrite = force_overwrite, **kwargs)
-
-
-def starmap(map_id: str, func: Callable, args, kwargs, force_overwrite: bool = False) -> result.MapResult:
-    return mapper.htmap(func).starmap(map_id, args, kwargs, force_overwrite = force_overwrite)
-
-
-def build_map(map_id: str, func: Callable, force_overwrite: bool = False) -> mapper.MapBuilder:
-    return mapper.htmap(func).build_map(map_id, force_overwrite = force_overwrite)
 
 
 def recover(map_id: str) -> result.MapResult:
@@ -26,3 +16,7 @@ def remove(map_id: str, not_exist_ok = True):
     except exceptions.MapIDNotFound as e:
         if not not_exist_ok:
             raise e
+
+
+def force_remove(map_id: str):
+    shutil.rmtree(mapper.map_dir_path(map_id))
